+++ src/lib/filter.test.ts
@@ -1,13 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { expect, test, describe, it } from "@jest/globals";
+import { expect, test, describe } from "@jest/globals";
 import { kind, modelToGroupVersionKind } from "kubernetes-fluent-client";
 import * as fc from "fast-check";
 import { CreatePod, DeletePod } from "../fixtures/loader";
 import { shouldSkipRequestRegex } from "./filter";
-import * as sut from "./filter";
-import { AdmissionRequest, Binding, DeepPartial, Event, Operation } from "./types";
+import { AdmissionRequest, Binding, Event } from "./types";
 
 export const callback = () => undefined;
 
@@ -652,453 +651,4 @@ test("should processing when deletionTimestamp is not present on pod", () => {
   };
 
   expect(shouldSkipRequestRegex(binding, pod, [])).toBe(false);
-});
-
-describe("definedEvent", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, ""],
-    [{ event: "" }, ""],
-    [{ event: "nonsense" }, "nonsense"],
-    [{ event: Event.Create }, Event.Create],
-    [{ event: Event.CreateOrUpdate }, Event.CreateOrUpdate],
-    [{ event: Event.Update }, Event.Update],
-    [{ event: Event.Delete }, Event.Delete],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedEvent(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definesDelete", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, false],
-    [{ event: "" }, false],
-    [{ event: "nonsense" }, false],
-    [{ event: Event.Create }, false],
-    [{ event: Event.CreateOrUpdate }, false],
-    [{ event: Event.Update }, false],
-    [{ event: Event.Delete }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definesDelete(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("misboundDeleteWithDeletionTimestamp", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, false],
-    [{ event: "" }, false],
-    [{ event: "nonsense" }, false],
-    [{ event: Event.Create }, false],
-    [{ event: Event.CreateOrUpdate }, false],
-    [{ event: Event.Update }, false],
-    [{ event: Event.Delete }, false],
-    [{ event: Event.Delete, filters: {} }, false],
-    [{ event: Event.Delete, filters: { deletionTimestamp: false } }, false],
-    [{ event: Event.Delete, filters: { deletionTimestamp: true } }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.misboundDeleteWithDeletionTimestamp(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("operationMatchesEvent", () => {
-  //[ Operation, Event, result ]
-  it.each([
-    ["", "", true],
-    ["", Event.Create, false],
-    [Operation.CREATE, "", false],
-
-    [Operation.CREATE, Event.Create, true],
-    [Operation.CREATE, Event.Update, false],
-    [Operation.CREATE, Event.Delete, false],
-    [Operation.CREATE, Event.CreateOrUpdate, true],
-    [Operation.CREATE, Event.Any, true],
-
-    [Operation.UPDATE, Event.Create, false],
-    [Operation.UPDATE, Event.Update, true],
-    [Operation.UPDATE, Event.Delete, false],
-    [Operation.UPDATE, Event.CreateOrUpdate, true],
-    [Operation.UPDATE, Event.Any, true],
-
-    [Operation.DELETE, Event.Create, false],
-    [Operation.DELETE, Event.Update, false],
-    [Operation.DELETE, Event.Delete, true],
-    [Operation.DELETE, Event.CreateOrUpdate, false],
-    [Operation.DELETE, Event.Any, true],
-
-    [Operation.CONNECT, Event.Create, false],
-    [Operation.CONNECT, Event.Update, false],
-    [Operation.CONNECT, Event.Delete, false],
-    [Operation.CONNECT, Event.CreateOrUpdate, false],
-    [Operation.CONNECT, Event.Any, true],
-  ])("given operation %s and event %s, returns %s", (op, evt, expected) => {
-    const result = sut.operationMatchesEvent(op, evt);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("declaredOperation", () => {
-  //[ AdmissionRequest, result ]
-  it.each([
-    [{}, ""],
-    [{ operation: null }, ""],
-    [{ operation: "" }, ""],
-    [{ operation: "operation" }, "operation"],
-    [{ operation: Operation.CONNECT }, Operation.CONNECT],
-    [{ operation: Operation.CREATE }, Operation.CREATE],
-    [{ operation: Operation.UPDATE }, Operation.UPDATE],
-    [{ operation: Operation.DELETE }, Operation.DELETE],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const request = given as DeepPartial<AdmissionRequest>;
-
-    const result = sut.declaredOperation(request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("mismatchedEvent", () => {
-  //[ Binding, AdmissionRequest, result ]
-  it.each([
-    [{}, {}, false],
-    [{}, { operation: Operation.CREATE }, true],
-    [{ event: Event.Create }, {}, true],
-
-    [{ event: Event.Create }, { operation: Operation.CREATE }, false],
-    [{ event: Event.Update }, { operation: Operation.CREATE }, true],
-    [{ event: Event.Delete }, { operation: Operation.CREATE }, true],
-    [{ event: Event.CreateOrUpdate }, { operation: Operation.CREATE }, false],
-    [{ event: Event.Any }, { operation: Operation.CREATE }, false],
-
-    [{ event: Event.Create }, { operation: Operation.UPDATE }, true],
-    [{ event: Event.Update }, { operation: Operation.UPDATE }, false],
-    [{ event: Event.Delete }, { operation: Operation.UPDATE }, true],
-    [{ event: Event.CreateOrUpdate }, { operation: Operation.UPDATE }, false],
-    [{ event: Event.Any }, { operation: Operation.UPDATE }, false],
-
-    [{ event: Event.Create }, { operation: Operation.DELETE }, true],
-    [{ event: Event.Update }, { operation: Operation.DELETE }, true],
-    [{ event: Event.Delete }, { operation: Operation.DELETE }, false],
-    [{ event: Event.CreateOrUpdate }, { operation: Operation.DELETE }, true],
-    [{ event: Event.Any }, { operation: Operation.DELETE }, false],
-
-    [{ event: Event.Create }, { operation: Operation.CONNECT }, true],
-    [{ event: Event.Update }, { operation: Operation.CONNECT }, true],
-    [{ event: Event.Delete }, { operation: Operation.CONNECT }, true],
-    [{ event: Event.CreateOrUpdate }, { operation: Operation.CONNECT }, true],
-    [{ event: Event.Any }, { operation: Operation.CONNECT }, false],
-  ])("given binding %j and admission request %j, returns %s", (bnd, req, expected) => {
-    const binding = bnd as DeepPartial<Binding>;
-    const request = req as DeepPartial<AdmissionRequest>;
-
-    const result = sut.mismatchedEvent(binding, request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definedGroup", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, ""],
-    [{ kind: null }, ""],
-    [{ kind: {} }, ""],
-    [{ kind: { group: null } }, ""],
-    [{ kind: { group: "" } }, ""],
-    [{ kind: { group: "group" } }, "group"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedGroup(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definesGroup", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, false],
-    [{ kind: null }, false],
-    [{ kind: {} }, false],
-    [{ kind: { group: null } }, false],
-    [{ kind: { group: "" } }, false],
-    [{ kind: { group: "group" } }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definesGroup(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("declaredGroup", () => {
-  //[ AdmissionRequest, result ]
-  it.each([
-    [{}, ""],
-    [{ kind: null }, ""],
-    [{ kind: {} }, ""],
-    [{ kind: { group: null } }, ""],
-    [{ kind: { group: "" } }, ""],
-    [{ kind: { group: "group" } }, "group"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const request = given as DeepPartial<AdmissionRequest>;
-
-    const result = sut.declaredGroup(request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("mismatchedGroup", () => {
-  //[ Binding, AdmissionRequest, result ]
-  it.each([
-    [{}, {}, false],
-    [{}, { kind: { group: "group" } }, false],
-    [{ kind: { group: "group" } }, {}, true],
-    [{ kind: { group: "group" } }, { kind: { group: "wrong" } }, true],
-    [{ kind: { group: "group" } }, { kind: { group: "group" } }, false],
-  ])("given binding %j and admission request %j, returns %s", (bnd, req, expected) => {
-    const binding = bnd as DeepPartial<Binding>;
-    const request = req as DeepPartial<AdmissionRequest>;
-
-    const result = sut.mismatchedGroup(binding, request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definedVersion", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, ""],
-    [{ kind: null }, ""],
-    [{ kind: {} }, ""],
-    [{ kind: { version: null } }, ""],
-    [{ kind: { version: "" } }, ""],
-    [{ kind: { version: "version" } }, "version"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedVersion(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definesVersion", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, false],
-    [{ kind: null }, false],
-    [{ kind: {} }, false],
-    [{ kind: { version: null } }, false],
-    [{ kind: { version: "" } }, false],
-    [{ kind: { version: "version" } }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definesVersion(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("declaredVersion", () => {
-  //[ AdmissionRequest, result ]
-  it.each([
-    [{}, ""],
-    [{ kind: null }, ""],
-    [{ kind: {} }, ""],
-    [{ kind: { version: null } }, ""],
-    [{ kind: { version: "" } }, ""],
-    [{ kind: { version: "version" } }, "version"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const request = given as DeepPartial<AdmissionRequest>;
-
-    const result = sut.declaredVersion(request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("mismatchedVersion", () => {
-  //[ Binding, AdmissionRequest, result ]
-  it.each([
-    [{}, {}, false],
-    [{}, { kind: { version: "version" } }, false],
-    [{ kind: { version: "version" } }, {}, true],
-    [{ kind: { version: "version" } }, { kind: { version: "wrong" } }, true],
-    [{ kind: { version: "version" } }, { kind: { version: "version" } }, false],
-  ])("given binding %j and admission request %j, returns %s", (bnd, req, expected) => {
-    const binding = bnd as DeepPartial<Binding>;
-    const request = req as DeepPartial<AdmissionRequest>;
-
-    const result = sut.mismatchedVersion(binding, request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definedKind", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, ""],
-    [{ kind: null }, ""],
-    [{ kind: {} }, ""],
-    [{ kind: { kind: null } }, ""],
-    [{ kind: { kind: "" } }, ""],
-    [{ kind: { kind: "kind" } }, "kind"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedKind(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definesKind", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, false],
-    [{ kind: null }, false],
-    [{ kind: {} }, false],
-    [{ kind: { kind: null } }, false],
-    [{ kind: { kind: "" } }, false],
-    [{ kind: { kind: "kind" } }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definesKind(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("declaredKind", () => {
-  //[ AdmissionRequest, result ]
-  it.each([
-    [{}, ""],
-    [{ kind: null }, ""],
-    [{ kind: {} }, ""],
-    [{ kind: { kind: null } }, ""],
-    [{ kind: { kind: "" } }, ""],
-    [{ kind: { kind: "kind" } }, "kind"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const request = given as DeepPartial<AdmissionRequest>;
-
-    const result = sut.declaredKind(request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("mismatchedKind", () => {
-  //[ Binding, AdmissionRequest, result ]
-  it.each([
-    [{}, {}, false],
-    [{}, { kind: { kind: "kind" } }, false],
-    [{ kind: { kind: "kind" } }, {}, true],
-    [{ kind: { kind: "kind" } }, { kind: { kind: "wrong" } }, true],
-    [{ kind: { kind: "kind" } }, { kind: { kind: "kind" } }, false],
-  ])("given binding %j and admission request %j, returns %s", (bnd, req, expected) => {
-    const binding = bnd as DeepPartial<Binding>;
-    const request = req as DeepPartial<AdmissionRequest>;
-
-    const result = sut.mismatchedKind(binding, request);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definedCategory", () => {
-  //[ Binding, result ]
-  it.each([
-    [{}, ""],
-    [{ isValidate: true }, "Validate"],
-    [{ isMutate: true }, "Mutate"],
-    [{ isWatch: true }, "Watch"],
-    [{ isFinalize: true, isWatch: true }, "Finalize"],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedCategory(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definedCallback", () => {
-  const validateCallback = () => {};
-  const mutateCallback = () => {};
-  const watchCallback = () => {};
-  const finalizeCallback = () => {};
-
-  //[ Binding, result ]
-  it.each([
-    [{}, null],
-    [{ isValidate: true, validateCallback }, validateCallback],
-    [{ isMutate: true, mutateCallback }, mutateCallback],
-    [{ isWatch: true, watchCallback }, watchCallback],
-    [{ isFinalize: true, finalizeCallback }, finalizeCallback],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedCallback(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("definedCallbackName", () => {
-  const validateCallback = () => {};
-  const mutateCallback = () => {};
-  const watchCallback = () => {};
-  const finalizeCallback = () => {};
-
-  //[ Binding, result ]
-  it.each([
-    [{}, ""],
-    [{ isValidate: true, validateCallback }, "validateCallback"],
-    [{ isMutate: true, mutateCallback }, "mutateCallback"],
-    [{ isWatch: true, watchCallback }, "watchCallback"],
-    [{ isFinalize: true, finalizeCallback }, "finalizeCallback"],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding = given as DeepPartial<Binding>;
-
-    const result = sut.definedCallbackName(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("declaredUid", () => {
-  //[ AdmissionRequest, result ]
-  it.each([
-    [{}, ""],
-    [{ uid: null }, ""],
-    [{ uid: "uid" }, "uid"],
-  ])("given %j, returns '%s'", (given, expected) => {
-    const request = given as DeepPartial<AdmissionRequest>;
-
-    const result = sut.declaredUid(request);
-
-    expect(result).toEqual(expected);
-  });
 });
